requirements2.txt
certifi==2020.12.5
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
gobject==0.1.0
google-pasta==0.1.7
grpcio==1.24.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.20.2
olefile==0.46
opencv-python==4.5.1.48
Pillow @ file:///tmp/build/80754af9/pillow_1617385572585/work
playsound==1.2.2
plyer==2.0.0
pycairo==1.20.0
pygame==2.0.1
PyGObject==3.40.1
PySimpleGUI==4.40.0
pyttsx3==2.90
six==1.15.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
